requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
requests==2.32.4
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
